+++ src/index.ts
@@ -1,18 +1,18 @@
-export { JettonWallet, JettonWalletOpCodes } from './ton/wrappers/JettonWallet';
-export type { JettonWalletData } from './ton/wrappers/JettonWallet';
 export { TacSdk } from './ton/sdk/TacSdk';
 export { TransactionStatus } from './ton/sdk/TransactionStatus';
 export { startTracking } from './ton/sdk/TxTracker';
+export * from './ton/sender';
 export type {
-    TacSDKTonClientParams,
-    JettonTransferData,
-    EvmProxyMsg,
-    TransactionLinker,
-    ShardMessage,
-    ShardTransaction,
     AssetBridgingData,
+    EvmProxyMsg,
     JettonBridgingData,
     JettonBurnData,
+    JettonTransferData,
+    ShardMessage,
+    ShardTransaction,
+    TacSDKTonClientParams,
+    TransactionLinker,
 } from './ton/structs/Struct';
 export { Network, SimplifiedStatuses } from './ton/structs/Struct';
-export * from './ton/sender';
+export type { JettonWalletData } from './ton/wrappers/JettonWallet';
+export { JettonWallet, JettonWalletOpCodes } from './ton/wrappers/JettonWallet';