+++ src/ton/sdk/TransactionStatus.ts
@@ -1,4 +1,5 @@
 import axios from 'axios';
+
 import type { TransactionLinker } from '../structs/Struct';
 import { SimplifiedStatuses } from '../structs/Struct';
 import { PUBLIC_LITE_SEQUENCER_ENDPOINTS } from './Consts';