+++ src/ton/wrappers/ContentUtils.ts
@@ -1,7 +1,7 @@
 // noinspection TypeScriptValidateTypes
 
-import { beginCell, Builder, Cell, Dictionary, Slice } from '@ton/core';
 import { Sha256 } from '@aws-crypto/sha256-js';
+import { beginCell, Builder, Cell, Dictionary, Slice } from '@ton/core';
 import axios from 'axios';
 import BN from 'bn.js';